ersky9x on companion9x (64bits) compilation

+++ src/simpgmspace.h
@@ -261,7 +261,7 @@ void StartEepromThread(const char *filename="eeprom.bin");
 extern const char *eepromFile;
 void eeprom_read_block (void *pointer_ram,
                    const void *pointer_eeprom,
-                        size_t size);
+                   uint32_t size);
 
 #define wdt_reset() sleep(1/*ms*/)
 #define board_init()
